requirements.txt
asgiref==3.6.0
camera==1.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
Flask==2.2.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
pyasn1==0.4.8
pycryptodome==3.18.0
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==2.0.3
Werkzeug==2.2.2
